Make key-agreement type buttons always sensitive.
Previously the user had to hit enter in the contact name entry when
creating a new contact. It seems that like is a little odd so this
changes makes the key-agreement type buttons always be enabled, but the
clicks will be ignored until an acceptable contact name has been
entered.

Fixes #12