Add a wrapper script in the bundle.
In order to avoid setting environment variables in the main binary (which was
racy), we now depend on pathrewrite to update the paths in the .dylibs. This
means that Pond needs to start in the correct directory so this change adds a
wrapper script that sets the working directory from the binary name.