Run PANDA scrypt processing in subprocess on 32-bit.
The GC is having issues on 32-bit systems (I think) and the biggest
allocation, by far, is the PANDA scrypt processing. This change causes
Pond to exec /proc/self/exec and feed the input to be hashes into a
subprocess. Thus that address space can be completed cleaned up.

We'll have to see whether this helps the reported issues.