Switch to late signing.
Previously, messages were signed for delivery (with the group signature)
as soon as they were enqueued. This resulted in a complex dance when
receiving a group update because messages in the outgoing queue needed
to be resigned.

This change switches to late signing: messages are only signed just
before transmission to the destination server.