Add commandline option to specify listen addr
That is the ip address of the interface to listen on
via Tor hidden service. This should usually be set to 127.0.0.1
but sometimes you want to set it to a bridge interface or something.

The ip addr is saved in the config just like the port is... and
then upon subsequent server starts it will be read from the config
instead of commandline.