Add the ability to entomb statefiles.
When TPM protection is enabled, it's not possible to move the statefile
from one computer to another. (Well, it's possible, but it won't work.)

In order to address this, this change adds support for "entombing"
statefiles: the statefile is written to a normal file, but encrypted
under a random key. The key is printed in the UI, but not otherwise
stored.

The user can then write the key down, move the entombed file to another
computer, import it under new TPM protection and then burn the paper.

(Not yet implemented in the CLI.)