Highlight contacts using old ratchet.
This code will be removed at some point so this change highlights
contacts with whom an old ratchet is being used. In order to ensure that
a contact keeps working, it's neccessary to confirm that the contact is
running an up-to-date version of the client and to perform a new key
exchange with them.