Add fingerprint of public keys
Added a fingerprint() method to both the client and Contact structs.
Display fingerprints along with User and Contact information.

Fingerprints should be safe to display publicly on sites like github
or twitter.  I donno if we should seed the fingerprinting though.

We avoid using the identity key in the fingerprint because it's known
by the server.