Use NewSimpleMeetingPlace in TestIntroductions
We need NewSimpleMeetingPlace to avoid hitting the PANDA server when
testing, but an intermitent race condition is created here.

I disabled the two calls to clients[1].ReloadWithMeetingPlace(mp) in
TestIntroductions because they break the client.gui.events channel
preventing the uiStatePANDAComplete message from arriving.

Appears ReloadWithMeetingPlace helps TestPANDA avoid a data race, but
the issue haasn't arisen for me.  If we ReloadWithMeetingPlace then
doGreet should be split into two seperate functions to actually press
greet and create the goroutine to wait for it.

I suspect however that launching the listening goroutine before
doing the greet on the other side avoids this data race better.