doc/index.html: update/fix Tails instructions
I just tested on Tails 1.3 and found that while the instructions result in a
working Pond (with GUI) they also result in a .bashrc that prints these two
errors in every new shell:

    unexpected EOF while looking for matching `''
    syntax error: unexpected end of file

One of the aliases (pond) was still set, but the others (pond-build, pond-cli,
pond-install-deps) are not set. I'm not sure how this could've happened as I do
remember testing the .bashrc before. Anyway, this commit fixes the instructions
so that the aliases get saved properly.